Introduce PageMarkerAnnotation
