+++ src/projects/suriano/Annotated/marker/SurianoMarker.tsx
@@ -1,6 +1,6 @@
 import { MarkerProps } from "../../../../components/Text/Annotated/core";
 import { MarkerBody } from "../../../../model/AnnoRepoAnnotation.ts";
-import { PageMarkerAnnotation } from "../../../../components/Text/Annotated/marker/DefaultMarkerAnnotation.tsx";
+import { PageMarkerAnnotation } from "../../../../components/Text/Annotated/marker/PageMarkerAnnotation.tsx";
 import { throwUnknownAnnotation } from "../../../../components/Text/Annotated/throwUnknownAnnotation.ts";
 import {
   projectPageMarkerAnnotationTypes,
